Refactor test setup functions, update comments
